Separated move and walk plugins
