+++ src/GameServer/MessageHandler/CharacterMoveHandlerPlugIn.cs
@@ -5,7 +5,6 @@
 namespace MUnique.OpenMU.GameServer.MessageHandler
 {
     using System.Runtime.InteropServices;
-    using MUnique.OpenMU.GameLogic.Views.World;
     using MUnique.OpenMU.Network.Packets.ClientToServer;
     using MUnique.OpenMU.Network.PlugIns;
     using MUnique.OpenMU.PlugIns;
@@ -20,8 +19,5 @@ namespace MUnique.OpenMU.GameServer.MessageHandler
     {
         /// <inheritdoc/>
         public override byte Key => InstantMoveRequest.Code;
-
-        /// <inheritdoc/>
-        public override MoveType MoveType => MoveType.Instant;
     }
 }
