+++ src/GameServer/MessageHandler/CharacterMoveHandlerPlugIn075.cs
@@ -5,7 +5,6 @@
 namespace MUnique.OpenMU.GameServer.MessageHandler
 {
     using System.Runtime.InteropServices;
-    using MUnique.OpenMU.GameLogic.Views.World;
     using MUnique.OpenMU.Network.PlugIns;
     using MUnique.OpenMU.PlugIns;
 
@@ -20,8 +19,5 @@ namespace MUnique.OpenMU.GameServer.MessageHandler
     {
         /// <inheritdoc/>
         public override byte Key => 0x11;
-
-        /// <inheritdoc/>
-        public override MoveType MoveType => MoveType.Instant;
     }
 }
